tweaked footer layout
