+++ src/components/Footer/BuildId.module.css
@@ -1,6 +1,7 @@
 .buildId {
-  display: inline-block;
+  display: block;
   font-size: var(--font-size-mini);
   margin-bottom: var(--spacer);
   font-family: var(--font-family-monospace);
+  text-align: center;
 }
