+++ src/components/Footer/Links.module.css
@@ -1,0 +1,3 @@
+.links {
+  text-align: center;
+}
